pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "open-knowledge-systems"
version = "0.1.0"
description = "Reference ETL slice for the Open Knowledge Systems repository."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Final State Press" }
]
dependencies = [
  "PyYAML>=6.0,<7.0"
]

[project.optional-dependencies]
graph = [
  "neo4j>=5.18,<6.0"
]

[project.scripts]
oks-etl = "oks.etl:run_cli"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
